Removed usage of BigDecimal. Fixed some errors. Fixed user book. Started work on MarketView.

+++ src/orderBookReconstructor/BuyOrder.java
@@ -1,12 +1,11 @@
 package orderBookReconstructor;
 
-import java.math.BigDecimal;
 import java.sql.Timestamp;
 
 import database.StockHandle;
 
 public class BuyOrder extends Order {
-	public BuyOrder(StockHandle stock, Timestamp timePlaced, BigDecimal price, int volume) {
+	public BuyOrder(StockHandle stock, Timestamp timePlaced, int price, int volume) {
 		super(stock, price, volume);
 	}
