+++ src/orderBookReconstructor/SellOrder.java
@@ -1,12 +1,11 @@
 package orderBookReconstructor;
 
-import java.math.BigDecimal;
 import java.sql.Timestamp;
 
 import database.StockHandle;
 
 public class SellOrder extends Order {
-	public SellOrder(StockHandle stock, Timestamp timePlaced, BigDecimal price, int volume) {
+	public SellOrder(StockHandle stock, Timestamp timePlaced, int price, int volume) {
 		super(stock, price, volume);
 	}
